requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-ads
Flask==2.2.2
langchain==0.0.134
openai==0.27.4
